setting frequency to double (bug fix)

+++ src/main/kotlin/DataClasses.kt
@@ -17,7 +17,7 @@ data class Customer(
     val lastPurchaseAt: String?,
     val totalSpending: Double,
     val averageOrderValue: Double,
-    val frequency: Int,
+    val frequency: Double,
     val preferredCategory: String?,
     val churned: Boolean?
 ) {
